pyproject.toml
[project]
name = "cells"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.10.0"
dependencies = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.8",
    "numpy>=1.26.4",
    "opencv-python~=4.10.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-javabridge>=4.0.4",
]
